Refactor: Move generic methods to TradeValidationUtils
