+++ core/src/test/java/bisq/core/trade/validation/TradeValidationTest.java
@@ -956,22 +956,22 @@ public class TradeValidationTest {
 
     @Test
     void checkValueInToleranceRejectsInvalidExpectedValueAndFactor() {
-        assertThrows(IllegalArgumentException.class, () -> TradeValidation.checkValueInTolerance(1, 0, 1));
-        assertThrows(IllegalArgumentException.class, () -> TradeValidation.checkValueInTolerance(1, -1, 1));
-        assertThrows(IllegalArgumentException.class, () -> TradeValidation.checkValueInTolerance(1, 1, 0.99));
+        assertThrows(IllegalArgumentException.class, () -> TradeValidationUtils.checkValueInTolerance(1, 0, 1));
+        assertThrows(IllegalArgumentException.class, () -> TradeValidationUtils.checkValueInTolerance(1, -1, 1));
+        assertThrows(IllegalArgumentException.class, () -> TradeValidationUtils.checkValueInTolerance(1, 1, 0.99));
     }
 
     @Test
     void checkByteArrayWithExpectedAcceptsMatchingByteArrays() {
         byte[] current = new byte[]{1, 2, 3};
 
-        assertSame(current, TradeValidation.checkByteArrayWithExpected(current, new byte[]{1, 2, 3}));
+        assertSame(current, TradeValidationUtils.checkByteArrayWithExpected(current, new byte[]{1, 2, 3}));
     }
 
     @Test
     void checkByteArrayWithExpectedRejectsMismatchingByteArrays() {
         IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> TradeValidation.checkByteArrayWithExpected(new byte[]{1, 2, 3}, new byte[]{1, 2, 4}));
+                () -> TradeValidationUtils.checkByteArrayWithExpected(new byte[]{1, 2, 3}, new byte[]{1, 2, 4}));
 
         assertEquals("current is not matching expected. current=010203, expected=010204", exception.getMessage());
     }
@@ -979,13 +979,13 @@ public class TradeValidationTest {
     @Test
     void checkByteArrayWithExpectedRejectsNullAndEmptyByteArrays() {
         assertThrows(NullPointerException.class,
-                () -> TradeValidation.checkByteArrayWithExpected(null, new byte[]{1}));
+                () -> TradeValidationUtils.checkByteArrayWithExpected(null, new byte[]{1}));
         assertThrows(NullPointerException.class,
-                () -> TradeValidation.checkByteArrayWithExpected(new byte[]{1}, null));
+                () -> TradeValidationUtils.checkByteArrayWithExpected(new byte[]{1}, null));
         assertThrows(IllegalArgumentException.class,
-                () -> TradeValidation.checkByteArrayWithExpected(new byte[0], new byte[]{1}));
+                () -> TradeValidationUtils.checkByteArrayWithExpected(new byte[0], new byte[]{1}));
         assertThrows(IllegalArgumentException.class,
-                () -> TradeValidation.checkByteArrayWithExpected(new byte[]{1}, new byte[0]));
+                () -> TradeValidationUtils.checkByteArrayWithExpected(new byte[]{1}, new byte[0]));
     }
 
     private static Offer offer(boolean isBuyOffer,
